requirements.txt
diffusers==0.31.0
accelerate==1.1.1
transformers==4.46.2
numpy==1.26.0
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
sentencepiece==0.2.0
SwissArmyTransformer==0.4.12
gradio==5.5.0
imageio==2.35.1
imageio-ffmpeg==0.5.1
openai==1.54.0
moviepy==1.0.3
scikit-video==1.1.11
spacy==3.7.6
einops==0.8.0
pydantic==2.9.1
opencv-python==4.10.0.84
matplotlib==3.9.2
markupsafe==2.1.5
h5py
scikit-image==0.24.0
pycocotools==2.0.8
gem_torch
open_clip_torch==2.24.0
decord
omegaconf
